Portfolio Page
Added lolcat and magic 8-ball projects to portfolio page. Added contact
information to bottom of page. Changed top navigation to add a "home"
link and a link back to the contact page on the home page. For some
reason, a bookmark to the contact section on this page would not work.